exclude test from race
